todo-application/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.27
psycopg[binary]==3.3.2
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]~=1.7.4
bcrypt==4.1.3
python-multipart==0.0.6
pydantic-settings==2.1.0
pytest==7.4.4
httpx==0.26.0
python-dotenv==1.0.0
